requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
python-dotenv==1.0.1
gunicorn==21.2.0
reportlab==4.2.2
openai>=1.40.0